Reworked template rendering and extension traits
Group extension traits and template rendering functionality by language
and put it all together in a module.  This allows the template code to
only see the extension traits for its language, which means we don't
need to prefix each method name with `webidl_`, `js_`, or `cpp_`.

I think this is a lot nicer, but one snag is that we still had naming
conflicts with the methods from the inherent impls (For example `name()`
and `namespace()` methods).  For now, I just tried to pick a name that
didn't conflict, like `nm()` instead of `name()`.  In the long-term, I
think the best solution would be to make the fields public on those
structs and remove the getter methods.  Then we could use those method
names in our extension traits.