Bug 1766045 - Initial work for UniFFI JS bindings
At this point, it can generate a WebIDL interface to handle the
scaffolding layer, which is a low-level C-ABI layer.  The next step is
to generate a JS wrapper on top of this that presents a nice API for
devs.

There's support for both synchronus and async functions, but we haven't
decided the how we want this to be configured.  In practice, almost all
functions will need to be async, so for now we're just forcing all
functions to be.

The `uniffi-bindgen-gecko-js` crate implements a UniFFI external
bindings generator and will generate the JS bindings. The `render`
module contains submodules that handling rendering different languages.
The basic system is that each module defines extension traits on the
`ComponentInterface` types (Type, Argument, Field, Function, etc.).  The
traits are then used by the template code to render things specific to
that language, for example how a function should be named in the WebIDL
file.  The `extend` library adds some syntactic sugar to allow this
without too much boilerplate.

The uniffi-js directory which contains shared code for
`uniffi-bindgen-gecko-js`.  Right now this is the `OwnedRustBuffer`
class, which is responsible for tracking ownership and freeing our
`RustBuffer` instances if needed.

UniFFIRustCallResult is a simply WebIDL dictionary that stores the
result of a scaffolding function call.

Differential Revision: https://phabricator.services.mozilla.com/D144472