Bug 1766045 - Adding xpcshell tests for our UniFFI fixtures
To test the bindings generation, we generate JS bindings for a set of test
Rust crates, then run xpcshell tests against those bindings.

Currently these tests only run when the `--enable-uniffi-fixtures`
option is set, so they need to be run manually be a dev.  I'd love to
create a system where these can optionally run in CI, but I'm not sure
how to do that.

Differential Revision: https://phabricator.services.mozilla.com/D144469