Use a static WebIDL interface
Changed the UniFFI WebIDL interface so that it's static rather than
generated.  Each function gets an id argument that the C++ code routes
to the correct scaffolding call.  The same system is also used for
reading/writing object pointers.

This allows us to drop all the generated .webidl and .h files and now
the generated C++ is consolidated into a single file.

The C++ code to implement the interface is still generated, but it's
basically a big switch that handles routing names to the C++ templates
in `uniffi-js`.

Using numeric IDs to identify functions/object pointers means that the
C++ code and JS code needs to agree on them.  In order to make that work
I changed `uniffi-bindgen-gecko-js` to input all UDL files at once,
calculate the ids, then generate both the C++ code and JS code using
those ids.